requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
sentence-transformers>=2.2.2
chromadb>=0.4.22
numpy>=1.25
scipy>=1.10
hnswlib>=0.8.0
click>=8.1.3
uvicorn>=0.29.0
fastapi>=0.110.0
pydantic>=1.10.0,<2.0  # chromadb needs v1, not v2
python-dotenv>=1.0.1
tiktoken==0.5.1
mysql-connector-python==8.2.0
Pillow==10.1.0
PyPDF2==3.0.1
python-docx==1.1.0
openai==1.3.0